feat: optimizing data types

+++ hopfield-neural-network/src/main/java/ua/kpi/iasa/sd/hopfieldneuralnetwork/controller/ImageProcessingController.java
@@ -17,7 +17,7 @@ public class ImageProcessingController {
     private final ImageProcessingService service;
 
     @PostMapping("/convert-to-pattern")
-    public int[][] processImages(@RequestParam("images") MultipartFile[] images) {
+    public byte[][] processImages(@RequestParam("images") MultipartFile[] images) {
         return service.processImages(images);
     }
